bookstore/auth-service/requirements.txt
Django==4.2.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
pika==1.3.2
requests==2.31.0
python-decouple==3.8
dj-database-url==2.1.0
django-cors-headers==4.3.1
gunicorn==21.2.0
PyJWT==2.8.0
bcrypt==4.1.2
